requirements.txt
# Core ML/RL libraries
absl-py==2.1.0
agentlace @ git+https://github.com/youliangtan/agentlace.git@cf2c337c5e3694cdbfc14831b239bd657bc4894d
annotated-types==0.7.0
array_record==0.6.0
astunparse==1.6.3
chex==0.1.88
distrax==0.1.5
einops==0.8.0
etils==1.11.0
flax==0.10.2
jax==0.4.35
jax-cuda12-pjrt==0.4.35
jax-cuda12-plugin==0.4.35
jaxlib==0.4.34
ml_collections==1.0.0
optax==0.2.4
opt_einsum==3.4.0
optree==0.13.1
orbax-checkpoint==0.10.3

# Gym environments
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.29.1
Farama-Notifications==0.0.4

# MuJoCo simulation
mujoco==2.3.7
dm-env==1.6
dm-tree==0.1.8
dm_robotics-transformations==0.8.1

# TensorFlow (for some utilities)
tensorflow==2.18.0
tensorflow-datasets==4.9.7
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.16.1
tensorflow-probability==0.25.0
tensorstore==0.1.71
tf_keras==2.18.0
keras==3.7.0
ml-dtypes==0.4.1

# Data processing
numpy==1.26.4
scipy==1.11.4
h5py==3.12.1
pyarrow==18.1.0
msgpack==1.1.0

# Computer vision
opencv-python==4.10.0.84
pillow==10.4.0
imageio==2.36.1
imageio-ffmpeg==0.5.1

# Visualization and logging
matplotlib==3.10.0
contourpy==1.3.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
wandb==0.19.1
tqdm==4.67.1

# Utilities
pynput==1.7.7
pyquaternion==0.9.9
PyYAML==6.0.2
requests==2.32.3
simple-parsing==0.1.6
moviepy==2.1.1
natsort==8.4.0
lxml==5.3.0

# Basic dependencies
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
distlib==0.3.9
distro==1.9.0
docstring_parser==0.16
docutils==0.21.2
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.55.3
fsspec==2024.12.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
glfw==2.8.0
google-pasta==0.2.0
grpcio==1.68.1
humanize==4.11.0
idna==3.10
immutabledict==4.2.1
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.5
kiwisolver==1.4.7
libclang==18.1.1
lz4==4.3.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
namex==0.0.8
nest-asyncio==1.6.0
nodeenv==1.9.1
packaging==24.2
platformdirs==4.3.6
pre-commit==3.3.3
proglog==0.1.10
promise==2.3
protobuf==3.20.3
psutil==6.1.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
PyOpenGL==3.1.7
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-xlib==0.33
pyzmq==26.2.0
rich==13.9.4
sentry-sdk==2.19.2
setproctitle==1.3.4
simplejson==3.19.3
six==1.17.0
smmap==5.0.1
termcolor==2.5.0
toml==0.10.2
toolz==1.0.0
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.28.0
Werkzeug==3.1.3
wrapt==1.17.0
zipp==3.21.0
zmq==0.0.0
